+++ AUTHORS
@@ -11,6 +11,7 @@ Development Lead
 Patches and Suggestions
 -----------------------
 
+- Carlos Henrique Guardão Gandarez <gandarez@gmail.com>
 - Patrick Auernig <dev.valeth@gmail.com>
 - James Lave <james@lave.cf>
 - Kieran Wynne <keewynne90@gmail.com>
