+++ package.json
@@ -19,7 +19,8 @@
     ],
     "homepage": "https://wakatime.com/",
     "bugs": {
-        "url": "https://github.com/wakatime/vscode-wakatime/issues"
+        "url": "https://github.com/wakatime/vscode-wakatime/issues",
+        "email": "support@wakatime.com"
     },
     "license": "BSD-3-Clause",
     "icon": "images/wakatime-logo.png",
@@ -35,16 +36,17 @@
         "url": "https://github.com/wakatime/vscode-wakatime.git"
     },
     "engines": {
-        "vscode": "^1.17.0",
+        "vscode": "^1.33.0",
         "sqlops": "*"
     },
     "main": "./dist/extension",
     "scripts": {
-        "setup": "mkdir -p ./dist/test",
+        "vscode:prepublish": "webpack --mode production",
+        "compile": "webpack --mode none",
+        "compile-dev": "tsc -p ./",
+        "watch": "webpack --mode none --watch",        
         "postinstall": "node ./node_modules/vscode/bin/install",
-        "webpack": "webpack --mode development",
-        "webpack-dev": "webpack --mode development --watch",
-        "vscode:prepublish": "webpack --mode production"
+        "test": "npm run compile-dev && node ./node_modules/vscode/bin/test"
     },
     "contributes": {
         "commands": [
@@ -71,15 +73,21 @@
         ]
     },
     "devDependencies": {
-        "@types/node": "^10.5.5",
-        "adm-zip": "0.4.11",
+        "@types/adm-zip": "^0.4.32",
+        "@types/mocha": "^5.2.6",
+        "@types/node": "^11.13.4",
+        "@types/request": "^2.48.1",
+        "@types/webpack": "^4.4.27",
+        "adm-zip": "0.4.13",
+        "del-cli": "^1.1.0",
         "original-fs": "^1.0.0",
-        "request": "2.87.0",
-        "rimraf": "2.6.2",
-        "ts-loader": "^4.4.2",
-        "typescript": "^3.0.3",
-        "vscode": "^1.1.22",
-        "webpack": "^4.19.1",
-        "webpack-cli": "^3.1.0"
+        "request": "2.88.0",
+        "rimraf": "^2.6.3",
+        "ts-loader": "^5.3.3",
+        "tslib": "^1.9.3",
+        "typescript": "^3.4.3",
+        "vscode": "^1.1.33",
+        "webpack": "^4.30.0",
+        "webpack-cli": "^3.3.0"
     }
 }
