+++ src/libs.ts
@@ -1,0 +1,6 @@
+export class Libs {
+    public static quote(str): string {
+        if (str.includes(' ')) return '"' + str.replace('"', '\\"') + '"';
+        return str;
+    }
+}
