+++ test/index.ts
@@ -16,7 +16,7 @@ var testRunner = require('vscode/lib/testrunner');
 // See https://github.com/mochajs/mocha/wiki/Using-mocha-programmatically#set-options for more info
 testRunner.configure({
 	ui: 'tdd', 		// the TDD UI is being used in extension.test.ts (suite, test, etc.)
-	useColors: true // colored output from test results
+	useColors: process.platform !== 'win32' // colored output from test results
 });
 
 module.exports = testRunner;
