+++ tsconfig.json
@@ -2,11 +2,25 @@
   "compilerOptions": {
     "module": "commonjs",
     "target": "es6",
-    "rootDir": "src",
     "outDir": "out",
-    "lib": ["es6"],
-    "sourceMap": true
+    "moduleResolution": "node",
+    "lib": [        
+      "es2018"
+    ],
+    "rootDir": ".",
+    "emitDecoratorMetadata": true,
+    "experimentalDecorators": true,
+    "noUnusedLocals": true,
+    "noUnusedParameters": true,
+    "strict": true,
+    "importHelpers": true,
+    "removeComments": true,
+    "noImplicitAny": false,
+    "strictPropertyInitialization": false,
   },
+  "include": [
+    "src/**/*"
+  ],
   "exclude": [
     "node_modules",
     "out/wakatime-master",
