+++ tslint.json
@@ -1,0 +1,7 @@
+{
+    "extends": "@smartive/tslint-config",
+    "rules": {
+        "import-name": false,
+        "ter-arrow-parens": false
+    }
+}
